+++ dinf/misc.py
@@ -84,13 +84,13 @@ Pytree = Any
 
 
 def _dtree_map(func, *trees: Pytree) -> Pytree:
-    return jax.tree_map(
+    return jax.tree_util.tree_map(
         func, *trees, is_leaf=lambda x: not isinstance(x, collections.abc.Mapping)
     )
 
 
 def _dtree_structure(tree: Pytree) -> Pytree:
-    return jax.tree_structure(_dtree_map(lambda _: (), tree))
+    return jax.tree_util.tree_structure(_dtree_map(lambda _: (), tree))
 
 
 def tree_equal(tree: Pytree, *others: Pytree) -> bool:
@@ -110,7 +110,7 @@ def tree_shape(tree: Pytree) -> Pytree:
     Return a pytree with the same dictionary structure as the given tree,
     but with non-dictionaries replaced by their shape.
     """
-    return jax.tree_map(
+    return jax.tree_util.tree_map(
         lambda x: np.shape(x),
         tree,
         is_leaf=lambda x: isinstance(x, (list, tuple)),
@@ -121,7 +121,7 @@ def tree_cons(a, tree: Pytree) -> Pytree:
     """
     Prepend ``a`` in all tuples of the given tree.
     """
-    return jax.tree_map(
+    return jax.tree_util.tree_map(
         lambda x: tuple((a,) + tuple(x)),
         tree,
         is_leaf=lambda x: isinstance(x, tuple),
@@ -132,19 +132,23 @@ def tree_car(tree: Pytree) -> Pytree:
     """
     Return a tree of the leading values of all tuples in the given tree.
     """
-    return jax.tree_map(lambda x: x[0], tree, is_leaf=lambda x: isinstance(x, tuple))
+    return jax.tree_util.tree_map(
+        lambda x: x[0], tree, is_leaf=lambda x: isinstance(x, tuple)
+    )
 
 
 def tree_cdr(tree: Pytree) -> Pytree:
     """
     Return a tree of the trailing values of all tuples in the given tree.
     """
-    return jax.tree_map(lambda x: x[1:], tree, is_leaf=lambda x: isinstance(x, tuple))
+    return jax.tree_util.tree_map(
+        lambda x: x[1:], tree, is_leaf=lambda x: isinstance(x, tuple)
+    )
 
 
 def leading_dim_size(tree: Pytree) -> int:
     """Size of the leading dimension (e.g. batch dimension) of each feature."""
-    sizes = np.array(jax.tree_flatten(tree_car(tree_shape(tree)))[0])
+    sizes = np.array(jax.tree_util.tree_flatten(tree_car(tree_shape(tree)))[0])
     # All features should have the same size for the leading dimension.
     assert np.all(sizes[0] == sizes[1:])
     return sizes[0]